Remove parent command's flags from subcommand usage
Issue #42 introduced the ability for subcommands to print help output
that includes its parent's flags. Unfortunately, this duplicated some
flags, such as `-h`, `-v`, and `-t`, that are also considered to be a
part of the subcommand. This makes the usage output look a bit odd, as
these flags are printed twice. This commit removes that duplication; it
does not print help output for flags that are also part of the parent
command's flags.